Modified code hierarchy and code readability by adding library module
