+++ src-tauri/src/library/mod.rs
@@ -1,0 +1,6 @@
+pub mod commands;
+pub mod models;
+pub mod service;
+
+// Re-export command handlers
+pub use commands::*;
